Finished setting up calibrator
This finished the calibrator which now has modes to process noise and
amplitude, and spit out a tuning config.

Next steps will be to take this config and run it through a one euro
filter according to the JS repo for pitch pipe.

Probably makes sense to use the multi dimensional one euro rust filter
for this.