Convert to builder pattern to make process more directive
Previously I had ported over the logic more or less directly from the
original JS repo. This included calibration states. In practice though
this left for some very fragile code that could easily be used
incorrectly.

For instance, we had a lot of optional fields that got filled as we
progressed through the calibration. This models the numerous null
initiated fields in the JS version, which is frankly a terrible
practice.

This converts the logic over so we have discrete calibrators using
the builder pattern. When one calibration stage is complete it can be
transformed into the next calibration stage, ensuring we don't have
unnecessary option types, and progression through the calibration
process is much clearer