Add abstraction over multi-dimensional one euro filter
Adds a basic three axis abstraction over a multi-dimensional one euro
filter. This was done because the library being used separates the
filter from the state which is obnoxious.

It also stores the raw (unfiltered) values. We should see if this is a
requirement of one euro filters in general and fork if necessary